Refactor AirTableStorage into multiple mixins (#53)
In preperating for adding Reaction Roles
