+++ ledger_bot/storage/__init__.py
@@ -1,0 +1,5 @@
+"""Module for interacting with the AirTable database."""
+
+from . import airtable_storage
+
+AirtableStorage = airtable_storage.AirtableStorage
