+++ ledger_bot/storage/airtable_storage.py
@@ -1,0 +1,61 @@
+"""The various models used by ledger-bot."""
+
+import asyncio
+import logging
+from typing import Optional
+
+from .base_storage import BaseStorage
+from .bot_messages_mixin import BotMessagesMixin
+from .members_mixin import MembersMixin
+from .transactions_mixin import TransactionsMixin
+
+log = logging.getLogger(__name__)
+
+
+class AirtableStorage(BotMessagesMixin, TransactionsMixin, MembersMixin, BaseStorage):
+    """
+    A class to interface with AirTable.
+
+    Attributes
+    ----------
+    airtable_key : str
+        The authentication key to connect to the base
+    bot_id : str
+        The id of ledger-bot
+    users_url : str
+        The endpoint for the users table
+    wines_url : str
+        The endpoint for the wines table
+    bot_messages_url : str
+        The endpoint for the bot_messages table
+    auth_header :
+        The authentication header
+
+    Methods
+    -------
+    insert_member(record, session)
+        Creates a new member entry
+
+    get_or_add_member(member)
+        Either returns the record for the given member, or creates an entry and returns that
+
+    insert_transaction(record, session)
+        Creates a new transaction entry
+
+    """
+
+    def __init__(
+        self,
+        airtable_base: str,
+        airtable_key: str,
+        bot_id: Optional[str],
+    ):
+        self.airtable_key = airtable_key
+        self.bot_id = bot_id
+        self.members_url = f"https://api.airtable.com/v0/{airtable_base}/members"
+        self.wines_url = f"https://api.airtable.com/v0/{airtable_base}/wines"
+        self.bot_messages_url = (
+            f"https://api.airtable.com/v0/{airtable_base}/bot_messages"
+        )
+        self.auth_header = {"Authorization": f"Bearer {self.airtable_key}"}
+        self._semaphore = asyncio.Semaphore(5)
